requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Storage
boto3==1.29.7
minio==7.2.0

# Queue
redis==5.0.1
rq==1.15.1
celery==5.3.4

# PDF processing
PyMuPDF==1.23.8
pdfplumber==0.10.3
camelot-py[cv]==0.11.0
ocrmypdf==15.4.3
pytesseract==0.3.10

# Image processing
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.26.2
scikit-image==0.22.0

# Layout detection
layoutparser[paddlepaddle]==0.3.4
torch==2.1.1
transformers==4.35.2

# OCR
paddleocr==2.7.3

# NLP / NER
spacy==3.7.2

# Cloud OCR adapters
google-cloud-documentai==2.20.1
boto3==1.29.7
azure-ai-formrecognizer==3.3.2

# Search
elasticsearch==8.11.0

# Monitoring
prometheus-client==0.19.0

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2

